+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Ability to turn off Labels on uncontrolled Airports and uncontrolled Airports with arrivals and/or departures
 - Added button to close the preloader after an issue occurred ([#66](https://github.com/marvk/vatprism/issues/66))
+- Added tooltip for airline information like name, ICAO, callsign and country when hovering over a callsign
 
 ### Changed